number-plate-detector-api/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
openai==1.35.2
python-dotenv==1.0.1
pydantic==2.5.2
Pillow==10.2.0
httpx==0.26.0 